addressing comments and fixes and tests
