+++ server/src/main/java/org/opensearch/index/compositeindex/datacube/startree/utils/SequentialDocValuesIterator.java
@@ -29,11 +29,6 @@ public class SequentialDocValuesIterator {
     private final DocIdSetIterator docIdSetIterator;
 
     /**
-     * The value associated with the latest document.
-     */
-    private Long docValue;
-
-    /**
      * The id of the latest document.
      */
     private int docId = -1;
@@ -48,39 +43,12 @@ public class SequentialDocValuesIterator {
     }
 
     /**
-     * Returns the value associated with the latest document.
-     *
-     * @return the value associated with the latest document
-     */
-    public Long getDocValue() {
-        return docValue;
-    }
-
-    /**
-     * Sets the value associated with the latest document.
-     *
-     * @param docValue the value to be associated with the latest document
-     */
-    public void setDocValue(Long docValue) {
-        this.docValue = docValue;
-    }
-
-    /**
      * Returns the id of the latest document.
      *
      * @return the id of the latest document
      */
-    public int getDocId() {
+    int getDocId() {
         return docId;
-    }
-
-    /**
-     * Sets the id of the latest document.
-     *
-     * @param docId the ID of the latest document
-     */
-    public void setDocId(int docId) {
-        this.docId = docId;
     }
 
     /**
@@ -97,7 +65,7 @@ public class SequentialDocValuesIterator {
         if (docId >= currentDocId) {
             return docId;
         }
-        setDocId(this.docIdSetIterator.nextDoc());
+        docId = this.docIdSetIterator.nextDoc();
         return docId;
     }
 
@@ -113,12 +81,7 @@ public class SequentialDocValuesIterator {
             if (docId == DocIdSetIterator.NO_MORE_DOCS || docId != currentDocId) {
                 return null;
             }
-            if (docValue == null) {
-                docValue = sortedNumericDocValues.nextValue();
-            }
-            Long nextValue = docValue;
-            docValue = null;
-            return nextValue;
+            return sortedNumericDocValues.nextValue();
 
         } else {
             throw new IllegalStateException("Unsupported Iterator requested for SequentialDocValuesIterator");
